reworked code, added gallary to the Home page
