+++ pages/Home/index.js
@@ -1,0 +1,30 @@
+import styled from "styled-components";
+import { search, mapImages } from "../../components/lib/cloudinary";
+
+export async function getServerSideProps() {
+  const results = await search();
+
+  const { resources } = results;
+  const images = mapImages(resources);
+
+  return {
+    props: { images },
+  };
+}
+
+export default function Home({ images }) {
+  return (
+    <div>
+      <h1>This is the Map Page - Under Construction</h1>
+      {images.map((data) => {
+        console.log("images map: ", data);
+        return <StyledImage key={data.asset_id} src={data.image} />;
+      })}
+    </div>
+  );
+}
+const StyledImage = styled.img`
+  width: 15%;
+  height: 15%;
+  padding: 1rem;
+`;
